Rename read_store to fileio and compile it to a static lib
