+++ src/fileio/fileio.h
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-namespace Read_Store
+namespace FileIO
 {
     void store_char_vector(vector<char> &vec, string path);
     bool read_char_vector(vector<char> &vec, string path, int required_size);
