+++ src/pruning/CMakeLists.txt
@@ -15,6 +15,7 @@ set(HEADER_FILES
         )
 
 include_directories(../cube)
+include_directories(../fileio)
 
 set(SOURCE_FILES
         prune.cpp
@@ -26,4 +27,7 @@ set(SOURCE_FILES
         ep2Pruning.cpp
         )
 
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/../fileio ${CMAKE_CURRENT_BINARY_DIR}/fileio)
+
 add_library(Prune STATIC ${SOURCE_FILES} ${HEADER_FILES})
+target_link_libraries(Prune FileIO)
